\name{BayesRegMicrobiome-package}
\alias{BayesRegMicrobiome-package}
\alias{BayesRegMicrobiome}
\docType{package}
\title{
\packageTitle{BayesRegMicrobiome}
}
\description{
\packageDescription{BayesRegMicrobiome}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesRegMicrobiome}
\packageIndices{BayesRegMicrobiome}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BayesRegMicrobiome}

Maintainer: \packageMaintainer{BayesRegMicrobiome}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
